modify(projects): Change from ts to json
