+++ portfolio/app/data/interfaceProjects.ts
@@ -1,0 +1,8 @@
+export interface Project {
+  id: number;
+  name: string;
+  description: string;
+  src: string;
+  technology: string[];
+  link?: string;
+}
